requirements.txt
flask>=3.0
flask-cors>=4.0
langchain>=0.3
langchain-openai>=0.2
langchain-chroma>=0.1
langchain-community>=0.3
chromadb>=0.5
openai>=1.0
pdfplumber>=0.11
pypdf>=4.0
python-docx>=1.1
